+++ plugins/repository-azure/src/main/java/org/opensearch/repositories/azure/AzureStorageSettings.java
@@ -32,6 +32,12 @@
 
 package org.opensearch.repositories.azure;
 
+import com.azure.core.util.logging.ClientLogger;
+import com.azure.identity.ManagedIdentityCredentialBuilder;
+import com.azure.storage.blob.BlobServiceClientBuilder;
+import com.azure.storage.common.implementation.Constants;
+import com.azure.storage.common.implementation.connectionstring.StorageConnectionString;
+import com.azure.storage.common.implementation.connectionstring.StorageEndpoint;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.collect.MapBuilder;
 import org.opensearch.common.settings.SecureSetting;
@@ -45,13 +51,16 @@ import org.opensearch.core.common.Strings;
 import org.opensearch.core.common.settings.SecureString;
 
 import java.net.InetAddress;
+import java.net.URI;
 import java.net.UnknownHostException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
+import java.util.function.Function;
 
 final class AzureStorageSettings {
+    private final ClientLogger logger = new ClientLogger(AzureStorageSettings.class);
 
     // prefix for azure client settings
     private static final String AZURE_CLIENT_PREFIX_KEY = "azure.client.";
@@ -75,6 +84,18 @@ final class AzureStorageSettings {
         AZURE_CLIENT_PREFIX_KEY,
         "sas_token",
         key -> SecureSetting.secureString(key, null)
+    );
+
+    /** Azure token credentials such as Managed Identity */
+    public static final AffixSetting<String> TOKEN_CREDENTIAL_TYPE_SETTING = Setting.affixKeySetting(
+        AZURE_CLIENT_PREFIX_KEY,
+        "token_credential_type",
+        key -> Setting.simpleString(key, value -> {
+            if (value != null && value != "") {
+                TokenCredentialType.valueOfType(value);
+            }
+        }, Property.NodeScope),
+        () -> ACCOUNT_SETTING
     );
 
     /** max_retries: Number of retries in case of Azure errors. Defaults to 3 (RetryPolicy.DEFAULT_CLIENT_RETRY_COUNT). */
@@ -194,7 +215,9 @@ final class AzureStorageSettings {
     );
 
     private final String account;
-    private final String connectString;
+    private final String tokenCredentialType;
+    private final Function<BlobServiceClientBuilder, BlobServiceClientBuilder> clientBuilder;
+    private final StorageEndpoint storageEndpoint;
     private final String endpointSuffix;
     private final TimeValue timeout;
     private final int maxRetries;
@@ -208,7 +231,9 @@ final class AzureStorageSettings {
     // copy-constructor
     private AzureStorageSettings(
         String account,
-        String connectString,
+        String tokenCredentialType,
+        Function<BlobServiceClientBuilder, BlobServiceClientBuilder> clientBuilder,
+        StorageEndpoint storageEndpoint,
         String endpointSuffix,
         TimeValue timeout,
         int maxRetries,
@@ -220,7 +245,9 @@ final class AzureStorageSettings {
         ProxySettings proxySettings
     ) {
         this.account = account;
-        this.connectString = connectString;
+        this.tokenCredentialType = tokenCredentialType;
+        this.clientBuilder = clientBuilder;
+        this.storageEndpoint = storageEndpoint;
         this.endpointSuffix = endpointSuffix;
         this.timeout = timeout;
         this.maxRetries = maxRetries;
@@ -236,6 +263,7 @@ final class AzureStorageSettings {
         String account,
         String key,
         String sasToken,
+        String tokenCredentialType,
         String endpointSuffix,
         TimeValue timeout,
         int maxRetries,
@@ -246,7 +274,17 @@ final class AzureStorageSettings {
         ProxySettings proxySettings
     ) {
         this.account = account;
-        this.connectString = buildConnectString(account, key, sasToken, endpointSuffix);
+        this.tokenCredentialType = tokenCredentialType;
+        if (this.tokenCredentialType == null || this.tokenCredentialType.isEmpty()) {
+            String connectString = buildConnectString(account, key, sasToken, endpointSuffix);
+            this.clientBuilder = (builder) -> builder.connectionString(connectString);
+            this.storageEndpoint = getStorageEndpointFromConnectString(connectString);
+        } else {
+            StorageEndpoint storageEndpoint = buildStorageEndpoint(account, endpointSuffix);
+            this.clientBuilder = (builder) -> builder.credential(new ManagedIdentityCredentialBuilder().build())
+                .endpoint(storageEndpoint.getPrimaryUri());
+            this.storageEndpoint = storageEndpoint;
+        }
         this.endpointSuffix = endpointSuffix;
         this.timeout = timeout;
         this.maxRetries = maxRetries;
@@ -256,6 +294,14 @@ final class AzureStorageSettings {
         this.readTimeout = readTimeout;
         this.responseTimeout = responseTimeout;
         this.proxySettings = proxySettings;
+    }
+
+    public String getTokenCredentialType() {
+        return tokenCredentialType;
+    }
+
+    public StorageEndpoint getStorageEndpoint() {
+        return storageEndpoint;
     }
 
     public String getEndpointSuffix() {
@@ -274,8 +320,8 @@ final class AzureStorageSettings {
         return proxySettings;
     }
 
-    public String getConnectString() {
-        return connectString;
+    public String getAccount() {
+        return account;
     }
 
     private static String buildConnectString(String account, @Nullable String key, @Nullable String sasToken, String endpointSuffix) {
@@ -325,6 +371,7 @@ final class AzureStorageSettings {
         final StringBuilder sb = new StringBuilder("AzureStorageSettings{");
         sb.append("account='").append(account).append('\'');
         sb.append(", timeout=").append(timeout);
+        sb.append(", tokenCredentialType=").append(tokenCredentialType).append('\'');
         sb.append(", endpointSuffix='").append(endpointSuffix).append('\'');
         sb.append(", maxRetries=").append(maxRetries);
         sb.append(", proxySettings=").append(proxySettings != ProxySettings.NO_PROXY_SETTINGS ? "PROXY_SET" : "PROXY_NOT_SET");
@@ -370,6 +417,7 @@ final class AzureStorageSettings {
                 account.toString(),
                 key.toString(),
                 sasToken.toString(),
+                getValue(settings, clientName, TOKEN_CREDENTIAL_TYPE_SETTING),
                 getValue(settings, clientName, ENDPOINT_SUFFIX_SETTING),
                 getValue(settings, clientName, TIMEOUT_SETTING),
                 getValue(settings, clientName, MAX_RETRIES_SETTING),
@@ -430,7 +478,9 @@ final class AzureStorageSettings {
                 entry.getKey(),
                 new AzureStorageSettings(
                     entry.getValue().account,
-                    entry.getValue().connectString,
+                    entry.getValue().tokenCredentialType,
+                    entry.getValue().clientBuilder,
+                    entry.getValue().storageEndpoint,
                     entry.getValue().endpointSuffix,
                     entry.getValue().timeout,
                     entry.getValue().maxRetries,
@@ -444,5 +494,25 @@ final class AzureStorageSettings {
             );
         }
         return mapBuilder.immutableMap();
+    }
+
+    public BlobServiceClientBuilder configure(BlobServiceClientBuilder builder) {
+        return clientBuilder.apply(builder);
+    }
+
+    private StorageEndpoint buildStorageEndpoint(String account, String endpointSuffix) {
+        String tokenCredentialEndpointSuffix = endpointSuffix;
+        if (!Strings.hasText(tokenCredentialEndpointSuffix)) {
+            // Default to "core.windows.net".
+            tokenCredentialEndpointSuffix = Constants.ConnectionStringConstants.DEFAULT_DNS;
+        }
+        final URI primaryBlobEndpoint = URI.create("https://" + account + ".blob." + tokenCredentialEndpointSuffix);
+        final URI secondaryBlobEndpoint = URI.create("https://" + account + "-secondary.blob." + tokenCredentialEndpointSuffix);
+        return new StorageEndpoint(primaryBlobEndpoint, secondaryBlobEndpoint);
+    }
+
+    private StorageEndpoint getStorageEndpointFromConnectString(String connectionString) {
+        final StorageConnectionString storageConnectionString = StorageConnectionString.create(connectionString, logger);
+        return storageConnectionString.getBlobEndpoint();
     }
 }