+++ plugins/repository-azure/src/test/java/org/opensearch/repositories/azure/AzureStorageServiceTests.java
@@ -33,6 +33,7 @@
 package org.opensearch.repositories.azure;
 
 import com.azure.core.http.policy.HttpPipelinePolicy;
+import com.azure.identity.CredentialUnavailableException;
 import com.azure.storage.blob.BlobServiceClient;
 import com.azure.storage.common.policy.RequestRetryPolicy;
 import org.opensearch.common.settings.MockSecureSettings;
@@ -111,6 +112,112 @@ public class AzureStorageServiceTests extends OpenSearchTestCase {
             assertThat(client1.getAccountUrl(), equalTo("https://myaccount1.blob.my_endpoint_suffix"));
             final BlobServiceClient client2 = azureStorageService.client("azure2").v1();
             assertThat(client2.getAccountUrl(), equalTo("https://myaccount2.blob.core.windows.net"));
+        }
+    }
+
+    public void testCreateClientWithEndpointSuffixWhenManagedIdentityIsEnabled() throws IOException {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure clients without account key and sas token.
+        secureSettings.setString("azure.client.azure1.account", "myaccount1");
+        secureSettings.setString("azure.client.azure2.account", "myaccount2");
+
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            // Enabled managed identity for all clients
+            .put("azure.client.azure1.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .put("azure.client.azure2.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            // Defined an endpoint suffix for azure client 1 only.
+            .put("azure.client.azure1.endpoint_suffix", "my_endpoint_suffix")
+            .build();
+        try (AzureRepositoryPlugin plugin = pluginWithSettingsValidation(settings)) {
+            final AzureStorageService azureStorageService = plugin.azureStoreService;
+            // Expect azure client 1 to use the custom endpoint suffix
+            final BlobServiceClient client1 = azureStorageService.client("azure1").v1();
+            assertThat(client1.getAccountUrl(), equalTo("https://myaccount1.blob.my_endpoint_suffix"));
+            // Expect azure client 2 to use the default endpoint suffix
+            final BlobServiceClient client2 = azureStorageService.client("azure2").v1();
+            assertThat(client2.getAccountUrl(), equalTo("https://myaccount2.blob.core.windows.net"));
+        }
+    }
+
+    public void testCreateClientWithInvalidEndpointSuffix() {
+        final MockSecureSettings secureSettings1 = new MockSecureSettings();
+        secureSettings1.setString("azure.client.azure1.account", "myaccount1");
+
+        final MockSecureSettings secureSettings2 = new MockSecureSettings();
+        secureSettings2.setString("azure.client.azure2.account", "myaccount2");
+        secureSettings2.setString("azure.client.azure2.key", encodeKey("mykey12"));
+
+        final MockSecureSettings secureSettings3 = new MockSecureSettings();
+        secureSettings3.setString("azure.client.azure3.account", "myaccount1");
+        secureSettings3.setString("azure.client.azure3.sas_token", encodeKey("mysastoken"));
+
+        // Test invalid endpointsuffix when token credential is used
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings1)
+            .put("azure.client.azure1.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .put("azure.client.azure1.endpoint_suffix", "invalid endpoint suffix")
+            .build();
+
+        final IllegalArgumentException e1 = expectThrows(
+            IllegalArgumentException.class,
+            () -> storageServiceWithSettingsValidation(settings)
+        );
+
+        // Test invalid endpointsuffix when account key is used
+        final Settings settings2 = Settings.builder()
+            .setSecureSettings(secureSettings2)
+            // Defined an invalid endpoint suffix
+            .put("azure.client.azure2.endpoint_suffix", "invalid endpoint suffix")
+            .build();
+
+        final RuntimeException e2 = expectThrows(RuntimeException.class, () -> storageServiceWithSettingsValidation(settings2));
+
+        // Test invalid endpointsuffix when sas token key is used
+        final Settings settings3 = Settings.builder()
+            .setSecureSettings(secureSettings3)
+            // Defined an invalid endpoint suffix
+            .put("azure.client.azure3.endpoint_suffix", "invalid endpoint suffix")
+            .build();
+
+        final RuntimeException e3 = expectThrows(RuntimeException.class, () -> storageServiceWithSettingsValidation(settings3));
+    }
+
+    public void testGettingSecondaryStorageBlobEndpoint() throws IOException {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        secureSettings.setString("azure.client.azure1.account", "myaccount1");
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            // Enabled managed identity
+            .put("azure.client.azure1.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .build();
+        try (AzureRepositoryPlugin plugin = pluginWithSettingsValidation(settings)) {
+            final AzureStorageService azureStorageService = plugin.azureStoreService;
+            final Map<String, AzureStorageSettings> prevSettings = azureStorageService.refreshAndClearCache(Collections.emptyMap());
+            final Map<String, AzureStorageSettings> newSettings = AzureStorageSettings.overrideLocationMode(
+                prevSettings,
+                LocationMode.SECONDARY_ONLY
+            );
+            azureStorageService.refreshAndClearCache(newSettings);
+            final BlobServiceClient client1 = azureStorageService.client("azure1").v1();
+            assertThat(client1.getAccountUrl(), equalTo("https://myaccount1-secondary.blob.core.windows.net"));
+        }
+    }
+
+    public void testClientUsingManagedIdentity() throws IOException {
+        // Enabled managed identity
+        final Settings settings = Settings.builder()
+            .setSecureSettings(buildSecureSettings())
+            .put("azure.client.azure1.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .build();
+        try (AzureRepositoryPlugin plugin = pluginWithSettingsValidation(settings)) {
+            final AzureStorageService azureStorageService = plugin.azureStoreService;
+            final BlobServiceClient client1 = azureStorageService.client("azure1").v1();
+
+            // Expect the client to use managed identity for authentication, and it should fail because managed identity environment is not
+            // setup in the test
+            final CredentialUnavailableException e = expectThrows(CredentialUnavailableException.class, () -> client1.getAccountInfo());
+            assertThat(e.getMessage(), is("Managed Identity authentication is not available."));
         }
     }
 
@@ -418,6 +525,183 @@ public class AzureStorageServiceTests extends OpenSearchTestCase {
         assertThat(name, is("path/to/myfile"));
         name = blobNameFromUri(new URI("https://127.0.0.1/container/path/to/myfile"));
         assertThat(name, is("path/to/myfile"));
+    }
+
+    public void testSettingTokenCredentialForAuthenticationIsCaseInsensitive() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure client without account key or sas token.
+        secureSettings.setString("azure.client.azure.account", "myaccount");
+        secureSettings.setString("azure.client.azure2.account", "myaccount");
+
+        // Enabled Managed Identity in the settings using lower case and mixed case
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure.token_credential_type", "managed_identity")
+            .put("azure.client.azure2.token_credential_type", "managed_IDENTITY")
+            .build();
+
+        final AzureStorageService mock = storageServiceWithSettingsValidation(settings);
+        assertEquals(mock.storageSettings.get("azure").getTokenCredentialType(), "managed_identity");
+        assertEquals(mock.storageSettings.get("azure2").getTokenCredentialType(), "managed_IDENTITY");
+    }
+
+    public void testSettingTokenCredentialForAuthenticationWithAlternativeEnumValue() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure client without account key or sas token.
+        secureSettings.setString("azure.client.azure.account", "myaccount");
+
+        // Enabled Managed Identity in the settings using lower case
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure.token_credential_type", "managed")
+            .build();
+
+        final AzureStorageService mock = storageServiceWithSettingsValidation(settings);
+        assertEquals(mock.storageSettings.get("azure").getTokenCredentialType(), "managed");
+    }
+
+    public void testSettingUnsupportedTokenCredentialForAuthentication() {
+        final String unsupported_token_credential_type = "TOKEN_CREDENTIAL_TYPE_THAT_DOES_NOT_EXIST";
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+
+        // Azure client without account key or sas token.
+        secureSettings.setString("azure.client.azure.account", "myaccount");
+
+        // Enable the unsupported token credential type
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure.token_credential_type", unsupported_token_credential_type)
+            .build();
+
+        final IllegalArgumentException e = expectThrows(
+            IllegalArgumentException.class,
+            () -> storageServiceWithSettingsValidation(settings)
+        );
+        assertEquals(
+            "The token credential type '"
+                + unsupported_token_credential_type
+                + "' is unsupported, please use one of the following values: "
+                + String.join(", ", TokenCredentialType.getTokenCredentialTypes()),
+            e.getMessage()
+        );
+    }
+
+    public void testTokenCredentialAuthenticationOverridesOtherFormOfAuthentications() {
+        final String token_credential_type = TokenCredentialType.MANAGED_IDENTITY.name();
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure1 with account key
+        secureSettings.setString("azure.client.azure1.account", "myaccount1");
+        secureSettings.setString("azure.client.azure1.key", encodeKey("mykey"));
+
+        // Azure 2 with sas token
+        secureSettings.setString("azure.client.azure2.account", "myaccount2");
+        secureSettings.setString("azure.client.azure2.sas_token", encodeKey("mysastoken"));
+
+        // Azure 3 with account key and sas token
+        secureSettings.setString("azure.client.azure3.account", "myaccount3");
+        secureSettings.setString("azure.client.azure3.key", encodeKey("mykey"));
+        secureSettings.setString("azure.client.azure3.sas_token", encodeKey("mysastoken"));
+
+        // Azure 4 without sas token and account key
+        secureSettings.setString("azure.client.azure4.account", "myaccount4");
+
+        // Enable Managed Identity in all azure clients
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure1.token_credential_type", token_credential_type)
+            .put("azure.client.azure2.token_credential_type", token_credential_type)
+            .put("azure.client.azure3.token_credential_type", token_credential_type)
+            .put("azure.client.azure4.token_credential_type", token_credential_type)
+            .build();
+        final AzureStorageService mock = storageServiceWithSettingsValidation(settings);
+
+        // Expect token credential authentication is selected over account key or sas token.
+        assertEquals(token_credential_type, mock.storageSettings.get("azure1").getTokenCredentialType());
+        assertEquals(token_credential_type, mock.storageSettings.get("azure2").getTokenCredentialType());
+        assertEquals(token_credential_type, mock.storageSettings.get("azure3").getTokenCredentialType());
+        assertEquals(token_credential_type, mock.storageSettings.get("azure4").getTokenCredentialType());
+    }
+
+    public void testTokenCredentialWhenAccountIsNotProvided() {
+        // Setting with an account specified
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+
+        // Enabled Managed Identity in the settings
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .build();
+        final Exception e = expectThrows(Exception.class, () -> storageServiceWithSettingsValidation(settings));
+
+        // Expect failure due to missing account name
+        assertEquals(
+            "missing required setting [azure.client.azure.account] for setting [azure.client.azure.token_credential_type]",
+            e.getMessage()
+        );
+    }
+
+    public void testAuthenticationMethodNotProvided() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure client without account key and sas token.
+        secureSettings.setString("azure.client.azure.account", "myaccount");
+
+        // Disabled Managed Identity in the settings by default
+        final Settings settings = Settings.builder().setSecureSettings(secureSettings).build();
+        final SettingsException e = expectThrows(SettingsException.class, () -> storageServiceWithSettingsValidation(settings));
+
+        // Expect fall back to authentication via sas token or account key when token credential is not specified.
+        assertEquals("Neither a secret key nor a shared access token was set.", e.getMessage());
+    }
+
+    public void testSettingTokenCredentialTypeToBeEmpty() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure clients without account key and sas token.
+        secureSettings.setString("azure.client.azure1.account", "myaccount");
+        secureSettings.setString("azure.client.azure2.account", "myaccount");
+
+        // Disabled Managed Identity in the settings by default
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure1.token_credential_type", "")
+            .put("azure.client.azure2.token_credential_type", " ")
+            .build();
+        final SettingsException e = expectThrows(SettingsException.class, () -> storageServiceWithSettingsValidation(settings));
+
+        // Expect fall back to authentication via sas token or account key when token credential is not specified.
+        assertEquals("Neither a secret key nor a shared access token was set.", e.getMessage());
+    }
+
+    public void testManagedIdentityIsEnabled() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure client without account key or sas token.
+        secureSettings.setString("azure.client.azure.account", "myaccount");
+
+        // Enabled Managed Identity in the settings.
+        final Settings settings = Settings.builder()
+            .setSecureSettings(secureSettings)
+            .put("azure.client.azure.token_credential_type", TokenCredentialType.MANAGED_IDENTITY.name())
+            .build();
+
+        final AzureStorageService mock = storageServiceWithSettingsValidation(settings);
+        assertEquals(mock.storageSettings.get("azure").getTokenCredentialType(), TokenCredentialType.MANAGED_IDENTITY.name());
+    }
+
+    public void testNonTokenCredentialAuthenticationEnabled() {
+        final MockSecureSettings secureSettings = new MockSecureSettings();
+        // Azure client account key.
+        secureSettings.setString("azure.client.azure1.account", "myaccount1");
+        secureSettings.setString("azure.client.azure1.sas_token", encodeKey("mysastoken"));
+
+        // Azure client with sas token
+        secureSettings.setString("azure.client.azure2.account", "myaccount2");
+        secureSettings.setString("azure.client.azure2.key", encodeKey("mykey"));
+
+        final Settings settings = Settings.builder().setSecureSettings(secureSettings).build();
+        final AzureStorageService mock = storageServiceWithSettingsValidation(settings);
+
+        // Expect token credential is not enabled
+        assertEquals(mock.storageSettings.get("azure1").getTokenCredentialType(), "");
+        assertEquals(mock.storageSettings.get("azure2").getTokenCredentialType(), "");
     }
 
     private static MockSecureSettings buildSecureSettings() {